refactor: admin module
