+++ libs/frontend/modules/admin/src/graphql/admin.endpoints.graphql
