+++ libs/frontend/modules/admin/src/index.ts
@@ -1,3 +1,2 @@
-export * from './use-cases/admin.endpoints'
-export * from './use-cases/reset-data/ResetDataButton'
-export * from './use-cases/seed-base-types/SeedBaseTypesButton'
+export * from './store'
+export * from './use-cases'
