+++ libs/frontend/modules/admin/src/store/adminEndpoints.ts
@@ -1,4 +1,4 @@
-import { api as generatedApi } from './admin.endpoints.graphql.gen'
+import { api as generatedApi } from '../graphql/admin.endpoints.graphql.gen'
 
 export { generatedApi as adminEndpoints }
 export const { useResetDataMutation, useSeedBaseTypesMutation } = generatedApi
