+++ libs/frontend/modules/admin/src/use-cases/reset-data/ResetDataButton.tsx
@@ -1,8 +1,9 @@
 import { Button } from 'antd'
-import { useResetDataMutation } from '../admin.endpoints'
+import { useResetDataMutation } from '../../store'
 
 export const ResetDataButton = () => {
   const [resetData] = useResetDataMutation()
+  const onClick = () => resetData()
 
-  return <Button onClick={() => resetData()}>Reset Data</Button>
+  return <Button onClick={onClick}>Reset Data</Button>
 }
